read-api/requirements.txt
flask
pymongo==2.4
python-dateutil < 2.0
pytz
nose
